# [FEM.HyperElastic] Remove optimization based on type of matrix in StandardTetrahedralFEMForceField